refactoring: map_reduce_star_plus in separate file
